+++ js/main.js
@@ -1,3 +1,3 @@
 import './big-picture.js';
+import './photo-filter.js';
 import './user-photo.js';
-import './filter.js';
