+++ js/util.js
@@ -29,4 +29,12 @@ const isEscapeKey = (evt) => evt.key === 'Escape';
 
 const getRandomArrayEl = (element) => element[getRandomInteger(0, element.length - 1)];
 
-export { getRandomInteger, getRandomArrayEl, isEscapeKey, showAlert };
+const debounce = (callback, timeoutDelay) => {
+  let timeoutId;
+  return (...rest) => {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
+  };
+};
+
+export { getRandomInteger, getRandomArrayEl, isEscapeKey, showAlert, debounce };
